Rename Component name to AppTag
